Remove redundancies, clean code, use standards

+++ src/main/java/frc/robot/subsystems/wrist/WristIOTalonSRX.java
@@ -54,6 +54,6 @@ public class WristIOTalonSRX implements WristIO {
 
   @Override
   public double getDegrees() {
-    return (double) wristMotor.getSelectedSensorPosition();
+    return wristMotor.getSelectedSensorPosition();
   }
 }
